% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_annum.R
\name{as_annum}
\alias{as_annum}
\title{as_annum}
\usage{
as_annum(date, from = "09-01", sep = " ", format = "\%Y")
}
\arguments{
\item{date}{A list of date(s)}

\item{from}{The first day of an annum in the format "mm-dd"}

\item{sep}{A separator string. Defaults to " ".}

\item{format}{Format of the year. Defaults to "Y"}
}
\value{

}
\description{
as_annum
}
