% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.r
\name{augment}
\alias{augment}
\alias{augment.smam}
\title{Augment}
\usage{
\method{augment}{smam}(x, data = NULL, newdata = NULL, se_fit = FALSE, ...)
}
\arguments{
\item{x}{an object of type \code{smam}}

\item{...}{arguments for generic consistency.}
}
\value{
An augmented data frame with added fields:
\describe{
 \item{.fitted}{The fitted values of the regression model.}
}
}
\description{
Augment a smam model.
}
\details{
Returns a table with predicted fits from a SMA model.
}
\seealso{
\code{\link{predict.smam}}.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
