% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.r
\name{coef.smam}
\alias{coef.smam}
\title{Extract Coefficients from smam}
\usage{
\method{coef}{smam}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{...}{Additional arguments passed to or from other methods}
}
\value{
A named numeric vector of coefficients.
}
\description{
Extract Coefficients from smam
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
