% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smak.r
\name{smak-NEWS}
\alias{smak-NEWS}
\title{News for package 'smak':}
\description{
News for package \sQuote{smak}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\smak}{\CRANpkg{smak}}
}
\section{\smak{} Initial Version 0.1.0 (2026-09-22) }{

\itemize{
\item first CRAN release.
}
}

