% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind}
\alias{rbind}
\title{rbind}
\usage{
rbind(..., col_names = NULL, col_classes = NULL)
}
\arguments{
\item{...}{Objects of class data.frame}

\item{col_names}{Names of the columns of the returned dataframe.
Defaults to NULL, which would create new column names in the form "col_n". It could take a character vector.
Column names could be inherited from a specific dataframe by specifying col_names as an integer indicating which dataframe from those provided by it's order or
by specifying a dataframe object the column names would be inherited from.}

\item{col_classes}{Class of the column vectors in the returned dataframe.
Defaults to NULL, which would create classes naturally.
It could take a character vector of coersion functions (e.g. c(as.character, as.factor, as.Date2)).
Classes could be inherited from a specific dataframe by specifying col_names as an integer indicating which dataframe from those provided by it's order or
by specifying a dataframe object the classes would be inherited from.}
}
\value{
An object of class data.frame
}
\description{
rbind
}
