% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smak.r
\docType{package}
\name{smak}
\alias{smak-package}
\alias{smak}
\title{scalable model averagking kit}
\description{
Scalable Model Averaging.
}
\note{
This package is maintained as a hobby.
}
\section{Frequentist Model Averaging}{


We consider the basic problem of regressing a scalar \eqn{y} against a vector of
independent variables \eqn{\vec{x}}.
There is a long history of practitioners applying some form of \emph{variable selection} to select
a subset of the vector \eqn{\vec{x}} to use in the regression modeling,
as a way to avoid overfitting and improve generalization to out-of-sample data.
Besides the intractability of testing all \eqn{2^p} subsets of the \eqn{p} variables in \eqn{\vec{x}},
variable selection is often a precursor to questionable statistical practice.

As an alternative to variable selection, Hjort and Claeskens introduced 
\href{https://www.jstor.org/stable/30045339}{Frequentist Model Averaging}.
The idea behind FMA is to average together the regression coefficients
built on different subsets of the \eqn{\vec{x}}, with weights determined
typically by one of two methods, Mallows model averaging or Jackknife model averaging.
}

\section{Scalable Model Averaging}{


While FMA supplies a stronger theoretical basis than variable selection,
it does not solve the computational issues.
Enter _Scalable Model Averaging_. 
Under this technique one first computes the singular value decomposition of the
matrix of observed independent variables, \eqn{X}.
Then one performs \eqn{p} regressions of the \eqn{y} against each of the left singular values of the \eqn{X},
then performs model averaging on only those \eqn{p} regressions.
Zhu \emph{et al.} show how to do this, and prove that the performance is equivalent
to the computationally intractable version run on all \eqn{2^p} regressions.

This package implements that process of performing SVD then regression and
weight selection. The regressions produce model fits, but not estimates of
uncertainty of the coefficients or total model error. For that one would
have to rely on, for example, bootstrapping.
}

\section{Legal Mumbo Jumbo}{


smak is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\references{
Zhu, Rong, Haiying Wang, Xinyu Zhang, and Hua Liang. 2023. 
“A Scalable Frequentist Model Averaging Method.” 
Journal of Business & Economic Statistics 41 (4): 1228–37. 
\doi{10.1080/07350015.2022.2116442}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shabbychef/smak}
  \item Report bugs at \url{https://github.com/shabbychef/smak/issues}
}

}
\author{
Steven E. Pav \email{shabbychef@gmail.com}

\strong{Maintainer}: Steven E. Pav \email{shabbychef@gmail.com} (\href{https://orcid.org/0000-0002-4197-6195}{ORCID})

}
\keyword{package}
