% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smamfit.r
\name{smam.control}
\alias{smam.control}
\title{Auxiliary for Controlling SMA Fitting}
\usage{
smam.control(
  wide_pragma = c("SIRS", "SIRSu", "alpha", "fixed_k", "error"),
  tall_pragma = c("tolerance", "none"),
  alpha = 0.99,
  k = 100
)
}
\arguments{
\item{wide_pragma}{what to do if X has more columns than rows, see Details.}

\item{tall_pragma}{what to do if X has as many rows as columns or more, see
Details.}

\item{alpha}{the \eqn{alpha} for the \dQuote{alpha} wide pragma.}

\item{k}{the number of singular values kept for the \dQuote{fixed_k} wide
pragma.}
}
\value{
A list.
}
\description{
Auxiliary function for \code{smam} fitting.
}
\details{
The \code{control} argument of \code{smam} is by default
passed to the \code{control} argument of \code{smamfit}, which uses its
selements as arguments to \code{smam.control}: the latter provides
defaults and sanity checking.

The wide pragma controls how fitting deals with wide matrices \eqn{X}. 
\describe{
 \item{SIRS}{Implements the SIRS procedure of Zhu et al. on \eqn{X}.}
 \item{SIRSu}{Implements the SIRS procedure on the left singular values
 \eqn{U}.}
 \item{alpha}{Performs SVD, then keeps the singular values that account for
 at least \eqn{\alpha} of the total sum of singular values. Always keeps at
 least one column, however.}
 \item{fixed_k}{Keep a fixed number of singular values.}
 \item{error}{Throws an error.}
}

The tall pragma controls how fitting deals with tall matrices \eqn{X}.
\describe{
 \item{none}{Takes all singular values.}
 \item{tolerance}{Keeps singular values up to machine epsilon ratio to the
 largest singular value.}
}
}
\references{
Zhu, Rong, Haiying Wang, Xinyu Zhang, and Hua Liang. 2023. 
“A Scalable Frequentist Model Averaging Method.” 
Journal of Business & Economic Statistics 41 (4): 1228–37. 
\doi{10.1080/07350015.2022.2116442}.

Zhu, Liping, Li, Lexin, Li, Runze, and Zhu, Lixing. 2011.
“Model-Free Feature Screening for Ultrahigh Dimensional Data."
J Am Stat Assoc. 106 (496): 1464-1475.
\doi{10.1198/jasa.2011.tm10563}.
}
\seealso{
\code{\link{smamfit}}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
