% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smamfit.r
\name{smam}
\alias{smam}
\alias{predict.smam}
\alias{print.smam}
\title{Friendly interface to scalable model averaging regression.}
\usage{
smam(
  formula,
  data,
  weights = NULL,
  na.action = na.omit,
  method = c("jackknife", "mallows"),
  ...
)

\method{predict}{smam}(object, newdata, na.action = na.pass, ...)

\method{print}{smam}(x, ...)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{lm} is called.}

\item{weights}{an optional vector of weights, or the string or bare name of the
weights in the \code{data} for use in the fitting process. Set to \code{NULL}
for none.}

\item{na.action}{How to deal with missing values in \code{y}, 
\code{X}, \code{wt}.}

\item{method}{the method to be used; for fitting, currently only
    \code{method = "qr"} is supported; \code{method = "model.frame"} returns
    the model frame (the same as with \code{model = TRUE}, see below).}

\item{...}{other arguments.}

\item{object}{Object of class inheriting from \code{smam}.}

\item{newdata}{a \code{data.frame} from which we can extract a model
frame via the formula of the \code{object}.}

\item{x}{an object used to select a method.}
}
\value{
An object of class \code{smam}.
}
\description{
A user friendly interface to the scalable frequentist model averaging
regression procedure.
}
\details{
Performs the scalable model averaging procedure to estimate linear
regression of \eqn{y} against \eqn{X}.
We find vector of coefficients \eqn{\beta} such that \eqn{X\beta} is 
approximately equal to \eqn{y}.
}
\examples{

nobs <- 100
nfeat <- 5
set.seed(1234)
X <- matrix(rnorm(nobs * nfeat),ncol=nfeat)
beta <- rnorm(nfeat)
y <- rnorm(nrow(X), mean=X \%*\% beta, sd=1)
# now the pretty frontend
data <- cbind(data.frame(outcome=y),as.data.frame(X))

fmla <- outcome ~ V1 + V2 + V3 + V4 + V5
fitm <- smam(fmla,data)

# with weights
data$wts <- runif(nrow(data),min=1,max=2)
fitw <- smam(fmla,data,weights=wts)

}
\seealso{
\code{\link{smamfit}}.

\code{\link{smam}}, \code{\link{smamfit}}.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{fitting}
