% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smamfit.r
\name{smamfit}
\alias{smamfit}
\title{scalable frequentist model averaging regression fit.}
\usage{
smamfit(
  y,
  X,
  method = c("jackknife", "mallows"),
  wt = NULL,
  sigma2 = NULL,
  Xnames = NULL,
  return.U = FALSE,
  return.X = TRUE,
  return.fitted = TRUE,
  control = list(...),
  formula = NULL,
  ...
)
}
\arguments{
\item{y}{vector of response data, of length \eqn{n}.}

\item{X}{a \eqn{n \times p}{n x p} matrix of independent variables.}

\item{method}{the selection method for the weights, either Jackknife model
averaging or Mallows' model averaging.}

\item{wt}{an optional vector of length \eqn{n} giving the weight of each
observation. If given must be non-negative. Weights are replication weights.}

\item{sigma2}{an optional squared variance to be used in Mallows' model
averaging. If not given, estimated from the data.}

\item{Xnames}{an optional vector of the names of the columns of X. If null
or not given, defaults to \code{colnames(X)}. Used only for naming the
output.}

\item{return.U}{a Boolean for whether the matrix U should be returned in
the output.}

\item{return.X}{a Boolean for whether the matrix X should be returned in
the output.}

\item{return.fitted}{a Boolean for whether the yhat fitted values should be returned in
the output.}

\item{control}{a list of parameters for controlling the the fitting process.
This is passed to \code{\link{smam.control}}.}

\item{formula}{an optional formula which will be saved in the object for prediction step later.}
}
\value{
An object of class \code{smam}.
}
\description{
Fits the scalable model averaging model to a regression problem.
}
\details{
Finds a vector of coefficients \eqn{\beta} such that \eqn{X\beta} is 
approximately equal to \eqn{y}. Does this by computing a singular
value decomposition of \eqn{X}, then performs \eqn{p} linear regressions
of the column of the left singular vectors against \eqn{y}, then
takes a weighted combination of those. 

Supports both the Mallows Model Averaging and Jackknife Model Averaging
approaches.
}
\note{
smak is free software.
}
\examples{

nobs <- 100
nfeat <- 5
set.seed(1234)
X <- matrix(rnorm(nobs * nfeat),ncol=nfeat)
beta <- rnorm(nfeat)
eta <- X \%*\% beta
y <- rnorm(ncol(X), mean=X \%*\% beta, sd=1)

mod0 <- smamfit(y=y,X=X)
summary(mod0)
}
\references{
Zhu, Rong, Haiying Wang, Xinyu Zhang, and Hua Liang. 2023. 
“A Scalable Frequentist Model Averaging Method.” 
Journal of Business & Economic Statistics 41 (4): 1228–37. 
\doi{10.1080/07350015.2022.2116442}.
}
\seealso{
the friendly interface \code{\link{smam}}.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{fitting}
