% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.r
\name{augment.smam}
\alias{augment.smam}
\title{Augment}
\usage{
\method{augment}{smam}(x, data = NULL, newdata = NULL, se_fit = FALSE, ...)
}
\arguments{
\item{x}{an object of type \code{smam}}

\item{data}{A \link[base:data.frame]{base::data.frame} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the original
data that was used to produce the object \code{x}. Defaults to
\code{stats::model.frame(x)} so that \code{augment(my_fit)} returns the augmented
original data. \strong{Do not} pass new data to the \code{data} argument.
Augment will report information such as influence and cooks distance for
data passed to the \code{data} argument. These measures are only defined for
the original training data.}

\item{newdata}{A \code{\link[base:data.frame]{base::data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing all
the original predictors used to create \code{x}. Defaults to \code{NULL}, indicating
that nothing has been passed to \code{newdata}. If \code{newdata} is specified,
the \code{data} argument will be ignored.}

\item{se_fit}{Logical indicating whether or not a \code{.se.fit} column should be
added to the augmented output. For some models, this calculation can be
somewhat time-consuming. Defaults to \code{FALSE}.}

\item{...}{arguments for generic consistency.}
}
\value{
An augmented data frame with added fields:
\describe{
 \item{.fitted}{The fitted values of the regression model.}
}
}
\description{
Augment a smam model.
}
\details{
Returns a table with predicted fits from a SMA model.
}
\seealso{
\code{\link{predict.smam}}.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
