% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs.r
\name{sirs}
\alias{sirs}
\title{Model-Free Feature Screening}
\usage{
sirs(
  y,
  X,
  thresholding = c("hard", "soft"),
  N = NULL,
  d_prop = 1,
  ties_pragma = c("ignore", "half_value")
)
}
\arguments{
\item{y}{vector of response data, of length \eqn{n}.}

\item{X}{a \eqn{n \times p}{n x p} matrix of independent variables.}

\item{thresholding}{the method for performing thresholding of the
\eqn{\hat{w}} variable to select predictors. For soft thresholding we
oversample with noise; for hard thresholding we pick the top \eqn{N} values.}

\item{N}{the number of variables to select for hard thresholding. Defaults
to \eqn{n / log(n)}.}

\item{ties_pragma}{the method for dealing with ties in the \eqn{y} values.
For \sQuote{ignore} we ignore them and implement the classic value of
Zhu et. al.  For \sQuote{half_value} the indicator function returns one half
for tied values in the \eqn{y}.}

\item{dprop}{for soft thresholding controls the number of noise variables
added to the set. We add \eqn{dprop p} variables.}
}
\value{
A sorted vector of the indices of which columns of \code{X} would be
selected, with values in \eqn{1, 2, \ldots, p}.
}
\description{
Performs the model-free feature screening procedure of Zhu et al. for
regression problems with many features.
}
\details{
Performs a model-free screening procedure. 
First each column of \eqn{X} is z-scored to get zero mean unit variance
variables. Then the \eqn{\hat{w}} is computed for each of the columns of
\eqn{X} against \eqn{y}. The kth element of this vector is computed as
\deqn{\frac{1}{n(n-1)(n-2)}sum_{1\le j \le n}\left(\sum_{1\le i \le n}
X_{i,k} I(y_i < y_j)\right)^2.}
Then we select indices \eqn{k} based on the values of this \eqn{\hat{w}},
based on the thresholding:

\describe{
\item{hard}{Select a fixed number of variates, based on the ordering of
\eqn{\hat{w}}.}
\item{soft}{Variates are selected if they are larger than the maximum of
\eqn{\hat{w}} computed on purely noise variates.}
}
}
\examples{
set.seed(1234)
X <- matrix(rnorm(200*500),ncol=500)
beta <- 2 * round(rnorm(ncol(X), sd=0.2)) * runif(ncol(X))
sum(abs(beta) > 0)
true_vars <- which(abs(beta) > 0)
y <- rnorm(nrow(X), mean=X \%*\% beta)
soft_vals <- sirs(y, X, thresholding='soft')
all(soft_vals \%in\% true_vars)

hard_vals <- sirs(y, X, thresholding='hard')
all(true_vars \%in\% hard_vals)
}
\references{
Zhu, Liping, Li, Lexin, Li, Runze, and Zhu, Lixing. 2011.
“Model-Free Feature Screening for Ultrahigh Dimensional Data."
J Am Stat Assoc. 106 (496): 1464-1475.
\doi{10.1198/jasa.2011.tm10563}.
}
\seealso{
\code{\link{smamfit}}.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
