% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smak.r
\docType{package}
\name{smak}
\alias{smak-package}
\alias{smak}
\title{statistics concerning the Markowitz portfolio}
\description{
Inference on the Markowitz portfolio.
}
\note{
This package is maintained as a hobby.
}
\section{Markowitz Portfolio}{


foo.
}

\section{Legal Mumbo Jumbo}{


smak is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\references{
Brandt, Michael W. "Portfolio Choice Problems." Handbook of Financial
Econometrics 1 (2009): 269--336. 
\url{http://shr.receptidocs.ru/docs/5/4748/conv_1/file1.pdf#page=298}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shabbychef/smak}
  \item Report bugs at \url{https://github.com/shabbychef/smak/issues}
}

}
\author{
Steven E. Pav \email{shabbychef@gmail.com}

\strong{Maintainer}: Steven E. Pav \email{shabbychef@gmail.com} (\href{https://orcid.org/0000-0002-4197-6195}{ORCID})

}
\keyword{package}
