% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.r
\name{tidy.smam}
\alias{tidy.smam}
\title{Tidy}
\usage{
\method{tidy}{smam}(x, ...)
}
\arguments{
\item{x}{an object of type \code{smam}.}

\item{...}{arguments for generic consistency.}
}
\value{
A tidy \code{tibble::tibble()} with fields
\describe{
 \item{term}{The name of the estimated parameter. Betas typically come before gammas.}
 \item{estimate}{The estimated parameter.}
}
}
\description{
Tidy a smam model.
}
\details{
Returns a table with information on the fit coefficients
from the scalable model averaging model.
At the moment, no standard errors are computed.
}
\examples{

nobs <- 100
nfeat <- 5
set.seed(1234)
X <- matrix(rnorm(nobs * nfeat),ncol=nfeat)
beta <- rnorm(nfeat)
eta <- X \%*\% beta
y <- rnorm(ncol(X), mean=X \%*\% beta, sd=1)

mod0 <- smamfit(y=y,X=X)
print(tidy(mod0))

}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
